+++ packages/amplify-category-api/src/provider-utils/awscloudformation/base-api-stack.ts
@@ -120,7 +120,7 @@ export abstract class ContainersStack extends cdk.Stack {
     const { pipelineWithAwaiter } = this.pipeline({
       skipWait,
       service,
-      containersInfo,
+      containersInfo: containersInfo.filter(container => container.repository),
       gitHubSourceActionInfo,
     });
 
@@ -375,12 +375,10 @@ export abstract class ContainersStack extends cdk.Stack {
           secrets,
         });
 
-        if (build) {
-          containersInfo.push({
-            container,
-            repository,
-          });
-        }
+        containersInfo.push({
+          container,
+          repository,
+        });
 
         // TODO: should we use hostPort too? check network mode
         portMappings?.forEach(({ containerPort, protocol, hostPort }) => {