+++ packages/amplify-appsync-simulator/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.3.6](https://github.com/aws-amplify/amplify-cli/compare/amplify-appsync-simulator@2.3.5...amplify-appsync-simulator@2.3.6) (2022-01-20)
+
+**Note:** Version bump only for package amplify-appsync-simulator
+
+
+
+
+
 ## [2.3.5](https://github.com/aws-amplify/amplify-cli/compare/amplify-appsync-simulator@2.3.4...amplify-appsync-simulator@2.3.5) (2022-01-13)